pipeline/requirements.txt
pylint
pytest
pandas
requests
python-dotenv
